requirements.txt
streamlit==1.36.0
pandas==2.2.2
matplotlib==3.9.1
shap==0.46.0
scikit-learn==1.5.1
imbalanced-learn==0.12.3
xgboost==2.1.0
numpy==2.0.0
google-cloud-storage==2.10.0